[NETBEANS-161] Changed license header introducer to simple slash-star
The new Apache license header on all .java, .cpp, .c files (etc)
started with slash-star-star (the Javadoc introducer). While strictly
speaking this is acceptable, the header is not a Javadoc comment,
and no present or future Javadoc parser should attempt to regard it
as Javadoc comment. Therefore all such occurrences are changed to
simple slash-star introducer.