[NETBEANS-2523] Installer creates invalid Win Registry value
The WindowsRegistry_set32BitValue0() function was incorrectly using
the length of the Registry name (e.g. 'NoModify') as the the length of 
value. Since the value we are setting is of type DWORD the length is
indeed always exactly 4 bytes, regardless of OS bitness. 
Solution: sizeof(dword) will always return 4, but we might as well 
have hard-coded the 4.